spenny_trail5s/requirements.txt
pandas
yfinance
python-dotenv
dropbox
openai
pdfplumber
python-docx
pytesseract
Pillow
